+++ test/test_tarray2.jl
@@ -12,12 +12,12 @@ function Turing.randclass(urn::PolyaUrn)
   else
     println(weights)
   end
-  return Int64(c)::Int64
+  return Int(c)::Int
 end
 
 function f()
   urn = PolyaUrn(1.72)
-  classes = tzeros(Int64, 50)
+  classes = tzeros(Int, 50)
   for i in 1:50
     classes[i]  = randclass(urn)
     u = unique(classes[1:i])
